+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## x.y.z 2023-xx-yy
+## 23.12.0 2023-07-31
 ### PaymentSheet
 * [Added] Enable SEPA Debit and iDEAL for SetupIntents and PaymentIntents with setup_future_usage. Note: PaymentSheet doesn't display saved SEPA Debit payment methods yet.